"Fixed" camera rotation
Previously, the camera matched the rotation of the car on all its axis, this however meant that when the car rolled around after lets say a crash or a failed jump, the camera would roll with it, resulting in an unpleasant camera control. Now the cameras rotation only matches the cars y axis rotation, meaning whichever direction the car faces, the cameras base orientation faces the same way. Yet there is still an unwanted behavior, whenever the car has a z- or x-axis rotation greater than a specific value, its y-axis changes a bit with it, depending on the size of the value, meaning in some extreme rotation cases, the camera will have an unpleasant rotation for fractions of a second, still, this behavior has been deemed better than the previous one. The described behaviors can be selected in the inspector on the main camera object itself via the mouse controller script.